# backend/api/openapi.yaml
openapi: "3.1.0"
info: {title: Oh-My-AIHub API, version: "0.1.0", description: "Oh-My-AIHub 前后端契约的唯一来源（见 ADR）。覆盖全部 /api/** JSON 接口；/v1/**、/v1beta/** 外部模型协议入口只登记路径与认证。\n\n约定：金额一律为十进制字符串；时间为 RFC 3339；错误统一为 {\"error\":{\"code\",\"message\"}}；除 GET/HEAD/OPTIONS 外的 /api 请求须携带与 Host 同源的 Origin。\n\n扩展字段 x-access 声明路由门禁，由契约测试与路由表逐项对照：public 无需会话；session 需会话（允许首次改密前访问）；ready 需会话且已完成首次改密；admin 在 ready 基础上需管理员；gateway_key 凭网关 API Key。"}
servers:
  - url: /
    description: "同源部署：开发环境由 Vite 代理，Compose 环境由 Nginx 代理。"
...
paths:
...
  /api/keys:
    get:
      operationId: listAPIKeys
      tags: [keys]
      summary: "我的 API Key"
      x-access: ready
      security:
        - sessionCookie: []
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ApiKeyList"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
    post:
      operationId: createAPIKey
      tags: [keys]
      summary: "创建 API Key"
      x-access: ready
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApiKeyConfigRequest"}
      responses:
        '201':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ApiKeyWithSecret"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/keys/{keyID}:
    get:
      operationId: getAPIKey
      tags: [keys]
      summary: "API Key 详情"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/keyID"
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ApiKeyEnvelope"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
    patch:
      operationId: updateAPIKey
      tags: [keys]
      summary: "更新 API Key"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/keyID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApiKeyConfigRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ApiKeyEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
    delete:
      operationId: deleteAPIKey
      tags: [keys]
      summary: "删除 API Key"
      description: "DELETE 请求携带 JSON 请求体（expected_version）。"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/keyID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApiKeyVersionRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ApiKeyEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/keys/{keyID}/rotate:
    post:
      operationId: rotateAPIKey
      tags: [keys]
      summary: "轮换 API Key"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/keyID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApiKeyVersionRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ApiKeyWithSecret"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/keys/{keyID}/disable:
    post:
      operationId: disableAPIKey
      tags: [keys]
      summary: "停用 API Key"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/keyID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApiKeyVersionRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ApiKeyEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/keys/{keyID}/enable:
    post:
      operationId: enableAPIKey
      tags: [keys]
      summary: "启用 API Key"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/keyID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApiKeyVersionRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ApiKeyEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/keys/{keyID}/pool-members:
    post:
      operationId: addAPIKeyPoolMember
      tags: [keys]
      summary: "向模型池追加报价"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/keyID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApiKeyPoolMemberRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ApiKeyEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/calls:
    get:
      operationId: listGatewayCalls
      tags: [calls]
      summary: "调用记录"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          required: false
          description: "返回条数上限。"
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GatewayCallList"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/calls/{callID}:
    get:
      operationId: getGatewayCall
      tags: [calls]
      summary: "调用详情"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/callID"
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GatewayCallEnvelope"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
  /api/dashboard:
    get:
      operationId: getDashboard
      tags: [calls]
      summary: "使用看板"
      x-access: ready
      security:
        - sessionCookie: []
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Dashboard"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
...
  /v1/chat/completions:
    post:
      operationId: proxyOpenAIChat
      tags: [gateway]
      summary: OpenAI Chat Completions
      description: "外部模型协议入口：请求与响应正文按上游协议原样透传，本规范只登记路径与认证，不描述正文。不走会话 Cookie，也不做同源校验。"
      x-access: gateway_key
      security:
        - bearerKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, description: "协议原生请求正文。"}
      responses:
        default: {description: "协议原生响应（含流式）与协议原生错误。"}
  /v1/responses:
    post:
      operationId: proxyOpenAIResponses
      tags: [gateway]
      summary: OpenAI Responses
      description: "外部模型协议入口：请求与响应正文按上游协议原样透传，本规范只登记路径与认证，不描述正文。不走会话 Cookie，也不做同源校验。"
      x-access: gateway_key
      security:
        - bearerKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, description: "协议原生请求正文。"}
      responses:
        default: {description: "协议原生响应（含流式）与协议原生错误。"}
  /v1/messages:
    post:
      operationId: proxyAnthropicMessages
      tags: [gateway]
      summary: Anthropic Messages
      description: "外部模型协议入口：请求与响应正文按上游协议原样透传，本规范只登记路径与认证，不描述正文。不走会话 Cookie，也不做同源校验。"
      x-access: gateway_key
      security:
        - anthropicKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, description: "协议原生请求正文。"}
      responses:
        default: {description: "协议原生响应（含流式）与协议原生错误。"}
  /v1beta/models/{model}:
    post:
      operationId: proxyGemini
      tags: [gateway]
      summary: Google Gemini generateContent / streamGenerateContent
      description: "外部模型协议入口：请求与响应正文按上游协议原样透传，本规范只登记路径与认证，不描述正文。不走会话 Cookie，也不做同源校验。"
      x-access: gateway_key
      security:
        - geminiKey: []
      parameters:
        - $ref: "#/components/parameters/model"
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, description: "协议原生请求正文。"}
      responses:
        default: {description: "协议原生响应（含流式）与协议原生错误。"}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: oma_session
      description: "登录后由服务端设置的 HttpOnly 会话 Cookie。"
    bearerKey: {type: http, scheme: bearer, description: "平台 API Key：Authorization: Bearer <key>（OpenAI 协议）。"}
    anthropicKey:
      type: apiKey
      in: header
      name: x-api-key
      description: "平台 API Key（Anthropic 协议）。"
    geminiKey:
      type: apiKey
      in: header
      name: x-goog-api-key
      description: "平台 API Key（Gemini 协议）。"
  parameters:
...
    keyID:
      name: keyID
      in: path
      required: true
      description: "API Key ID。"
      schema: {type: string}
    callID:
      name: callID
      in: path
      required: true
      description: "调用 ID。"
      schema: {type: string}
...
    model:
      name: model
      in: path
      required: true
      description: "规范模型 ID 与动作，例如 `canonical/model:generateContent`。"
      schema: {type: string}
  responses:
    BadRequest:
      description: "请求格式无效或查询参数非法（invalid_json、invalid_request、invalid_input）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Unauthorized:
      description: "未登录或会话失效（authentication_required）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Forbidden:
      description: "被拒绝：跨站请求（cross_origin_request）、首次登录须改密（password_change_required）、需要管理员（administrator_required）或无权操作（forbidden）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    NotFound:
      description: "资源不存在（not_found）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Conflict:
      description: "版本或状态冲突（conflict、ledger_conflict、c2c_conflict、fee_rate_conflict 等）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Unprocessable:
      description: "业务校验失败（invalid_input、insufficient_spendable_capacity、unsafe_upstream 等）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...
  schemas:
    Amount: {type: string, pattern: "^-?[0-9]+(\\.[0-9]{1,9})?$", description: "积分金额的十进制字符串，最多 9 位小数，不使用浮点数。"}
    Timestamp: {type: string, format: date-time, description: "RFC 3339 时间。"}
...
    Error:
      type: object
      properties:
        code: {type: string}
        message: {type: string}
      required: [code, message]
      additionalProperties: false
    ErrorResponse:
      type: object
      description: "所有 /api 错误的统一信封；code 为稳定的机器可读标识，message 为中文提示。"
      properties:
        error: {$ref: "#/components/schemas/Error"}
      required: [error]
      additionalProperties: false
...
    ApiKeyPool:
      type: object
      properties:
        id: {type: string}
        model_id: {type: string}
        model_name: {type: string}
        protocol: {type: string, enum: [openai_chat_completions, openai_responses, anthropic_messages, google_gemini_generate_content]}
        version: {type: integer}
        members:
          type: array
          items: {$ref: "#/components/schemas/ApiKeyPoolMember"}
        created_at: {$ref: "#/components/schemas/Timestamp"}
        updated_at: {$ref: "#/components/schemas/Timestamp"}
      required: [id, model_id, model_name, protocol, version, members, created_at, updated_at]
      additionalProperties: false
    ApiKey:
      type: object
      properties:
        id: {type: string}
        display_name: {type: string}
        prefix: {type: string}
        generation: {type: integer}
        status: {type: string, enum: [active, disabled, deleted]}
        version: {type: integer}
        pools:
          type: array
          items: {$ref: "#/components/schemas/ApiKeyPool"}
        last_used_at:
          anyOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        created_at: {$ref: "#/components/schemas/Timestamp"}
        updated_at: {$ref: "#/components/schemas/Timestamp"}
      required: [id, display_name, prefix, generation, status, version, pools, last_used_at, created_at, updated_at]
      additionalProperties: false
    ApiKeyEnvelope:
      type: object
      properties:
        key: {$ref: "#/components/schemas/ApiKey"}
      required: [key]
      additionalProperties: false
    ApiKeyList:
      type: object
      properties:
        keys:
          type: array
          items: {$ref: "#/components/schemas/ApiKey"}
      required: [keys]
      additionalProperties: false
    ApiKeyWithSecret:
      type: object
      properties:
        key: {$ref: "#/components/schemas/ApiKey"}
        secret: {type: string, description: "平台 API Key 明文，仅在创建或轮换时返回一次。"}
      required: [key, secret]
      additionalProperties: false
    ApiPoolRequest:
      type: object
      properties:
        model_id: {type: string}
        protocol: {type: string, enum: [openai_chat_completions, openai_responses, anthropic_messages, google_gemini_generate_content]}
        offer_ids:
          type: array
          items: {type: string}
      required: [model_id, protocol, offer_ids]
      additionalProperties: false
    ApiKeyConfigRequest:
      type: object
      properties:
        display_name: {type: string}
        expected_version: {type: integer, description: "仅更新时使用。"}
        pools:
          type: array
          items: {$ref: "#/components/schemas/ApiPoolRequest"}
      required: [display_name, pools]
      additionalProperties: false
    ApiKeyVersionRequest:
      type: object
      properties:
        expected_version: {type: integer}
      required: [expected_version]
      additionalProperties: false
    ApiKeyPoolMemberRequest:
      type: object
      properties:
        expected_version: {type: integer}
        model_id: {type: string}
        protocol: {type: string, enum: [openai_chat_completions, openai_responses, anthropic_messages, google_gemini_generate_content]}
        offer_id: {type: string}
        priority: {type: integer}
      required: [expected_version, model_id, protocol, offer_id]
      additionalProperties: false
...
    GatewayCall:
      type: object
      properties:
        id: {type: string}
        consumer_account_id: {type: string}
        api_key_id: {type: string}
        key_prefix: {type: string}
        key_generation: {type: integer}
        pool_id: {type: string}
        pool_version: {type: integer}
        model_id: {type: string}
        protocol: {type: string, enum: [openai_chat_completions, openai_responses, anthropic_messages, google_gemini_generate_content]}
        status: {type: string, enum: [rejected, in_progress, pending_delivery, succeeded, failed, incomplete, cancelled]}
        decision_code: {type: string}
        candidate_count: {type: integer}
        attempt_count: {type: integer}
        hold_id: {type: string}
        preauthorized: {$ref: "#/components/schemas/Amount"}
        zero_hold_reason: {type: string}
        fee_rate_version: {type: integer}
        fee_rate_nano: {type: integer}
        final_offer_id: {type: string}
        final_channel_name: {type: string}
        completion_reason: {type: string}
        usage:
          anyOf:
            - $ref: "#/components/schemas/Usage"
            - type: "null"
        provider_charge: {$ref: "#/components/schemas/Amount"}
        platform_fee: {$ref: "#/components/schemas/Amount"}
        settled_price_tier_seq: {type: integer}
        final_http_status: {type: integer}
        attempts:
          type: array
          items: {$ref: "#/components/schemas/GatewayAttempt"}
        created_at: {$ref: "#/components/schemas/Timestamp"}
        completed_at:
          anyOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
      required: [id, consumer_account_id, api_key_id, key_prefix, key_generation, pool_id, pool_version, model_id, protocol, status, decision_code, candidate_count, attempt_count, hold_id, preauthorized, zero_hold_reason, fee_rate_version, fee_rate_nano, final_offer_id, final_channel_name, completion_reason, usage, provider_charge, platform_fee, settled_price_tier_seq, final_http_status, attempts, created_at, completed_at]
      additionalProperties: false
    GatewayCallEnvelope:
      type: object
      properties:
        call: {$ref: "#/components/schemas/GatewayCall"}
      required: [call]
      additionalProperties: false
    GatewayCallList:
      type: object
      properties:
        calls:
          type: array
          items: {$ref: "#/components/schemas/GatewayCall"}
      required: [calls]
      additionalProperties: false
...
    Dashboard:
      type: object
      properties:
        consumer_spent: {$ref: "#/components/schemas/Amount"}
        provider_income: {$ref: "#/components/schemas/Amount"}
        today_spent: {$ref: "#/components/schemas/Amount"}
        today_succeeded_calls: {type: integer}
        today_external_provider_income: {$ref: "#/components/schemas/Amount"}
        active_key_count: {type: integer}
        pool_count: {type: integer}
        healthy_offer_count: {type: integer}
        unhealthy_offer_count: {type: integer}
        recent_calls:
          type: array
          items: {$ref: "#/components/schemas/GatewayCall"}
      required: [consumer_spent, provider_income, today_spent, today_succeeded_calls, today_external_provider_income, active_key_count, pool_count, healthy_offer_count, unhealthy_offer_count, recent_calls]
      additionalProperties: false
...